Use cases are validated

+++ src/main/java/com/keax/application/usecases/Institution/DeleteInstitutionUseCaseImpl.java
@@ -1,5 +1,6 @@
 package com.keax.application.usecases.Institution;
 
+import com.keax.domain.exceptions.ExceptionAlert;
 import com.keax.domain.ports.in.Institution.DeleteInstitutionUseCase;
 import com.keax.domain.ports.out.InstitutionRepositoryPort;
 
@@ -13,7 +14,11 @@ public class DeleteInstitutionUseCaseImpl implements DeleteInstitutionUseCase {
 
     @Override
     public Boolean deleteInstitution(Long institution_id) {
-        return institutionRepositoryPort.deleteInstitution(institution_id);
+        if (institutionRepositoryPort.existsByInstitutionIdAndInstitutionDeleted(institution_id, false)){
+            return institutionRepositoryPort.deleteInstitution(institution_id);
+        }
+
+        throw new ExceptionAlert("No existe la institución a eliminar.");
     }
 
 }
